fastapi_demo/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi-debug-toolbar>=0.6.3",
    "fastapi[all]>=0.119.1",
    "pydantic>=2.12.3",
    "uvicorn[standard]>=0.38.0",
]
